seperate plot counts/distribution
